Switch Appinsights for the web basic SDK
