+++ src/node/tsconfig.json
@@ -6,7 +6,6 @@
 		"module": "CommonJS",
 		"moduleResolution": "node",
 		"outDir": "../../dist/node",
-		"typeRoots": []
 	},
 	"include": [".", "../common", "../../vscode.proposed.telemetryLogger.d.ts"]
 }
